The wayfinder shows direction using programmer art now.

+++ src/main/scala/mod/iceandshadow3/util/SIntBits.java
@@ -1,29 +1,28 @@
 package mod.iceandshadow3.util;
 
-/** Utility functions for using the bits (esp. the sign bit) on an integer for
+/** Utility functions for using the bits (esp. the sign bit) on a long integer for
  * something else.
  */
 public class SIntBits {
-	public static boolean areAllSet(int value, int bitmask) {
+	public static boolean areAllSet(long value, long bitmask) {
 		return (value & bitmask) == bitmask;
 	}
 
-	public static boolean areAnySet(int value, int bitmask) {
+	public static boolean areAnySet(long value, long bitmask) {
 		return (value & bitmask) != 0;
 	}
 
-	public static int encode(int value, boolean flag) {
+	public static long encode(long value, boolean flag) {
 		return ((value & (-1 >>> 1)) + (flag ? 1 : 0)) * (flag ? -1 : 1);
 	}
 
-	// The following are common and simple-enough operations, but they're annoying
-	// to type out.
+	// The following are common and simple-enough operations, but are unintuitive or annoying.
 
-	public static boolean signBit(int value) {
+	public static boolean signBit(long value) {
 		return value < 0;
 	}
 
-	public static int value(int value) {
+	public static long value(long value) {
 		return Math.abs(value + (signBit(value) ? 1 : 0));
 	}
 }
